// XincaPro/src/app/Shared/models/perfilMod.ts
export class PerfilMod{
    IdPerfil: number;
    Nickname: string;
    Experiencia: number;
    Contrasena:string;
    Celular: number;
    Correo: string;
    IdCategoria: number;
    IdUsuario: number;
    NombreUsuario: string;
}

export class PerfilInsert{
    nickname: string;
    experiencia: number;
    celular: number;
    correo: string;
    IdCategoria: number;
    nombreUsuario: string; 
    contrasena: string;
}